# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maximizing Your Chat with AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update a user's status or role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin or attempting to demote last admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user permanently (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Cannot delete yourself or last admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/ai-usage:
    get:
      operationId: adminListAiUsage
      tags: [admin]
      summary: AI token usage per super admin for the current billing period (admin only)
      responses:
        "200":
          description: Per-tenant token usage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiUsageSummary"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/pricing:
    get:
      operationId: adminGetPricing
      tags: [admin]
      summary: Get the global usage-based pricing config (admin only)
      responses:
        "200":
          description: Current pricing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingConfig"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: adminUpdatePricing
      tags: [admin]
      summary: Update the global usage-based pricing config (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePricingInput"
      responses:
        "200":
          description: Updated pricing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingConfig"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/billing:
    get:
      operationId: adminListBilling
      tags: [admin]
      summary: Per-tenant subscription, usage and computed monthly bill (admin only)
      responses:
        "200":
          description: Per-tenant billing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminTenantBilling"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/revenue:
    get:
      operationId: adminGetRevenue
      tags: [admin]
      summary: Platform-wide MRR/ARR, tenant counts and spend trend (admin only)
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 365
          description: Trend window length in days (default 30).
      responses:
        "200":
          description: Revenue summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueSummary"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/subscriptions/{userId}:
    patch:
      operationId: adminRenewSubscription
      tags: [admin]
      summary: Renew, suspend or change a tenant's subscription (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: The tenant owner's user id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewSubscriptionInput"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionInfo"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    AdminUser:
      type: object
      required: [id, email, role, status, createdAt, approvedAt, ownerPhone]
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [pending, active, disabled]
        createdAt:
          type: string
          format: date-time
        approvedAt:
          type: ["string", "null"]
          format: date-time
        ownerPhone:
          type: ["string", "null"]
          description: Linked WhatsApp number, if the user has paired a device.
    AdminUpdateUserInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, disabled]
        role:
          type: string
          enum: [user, admin]
    AiUsageSummary:
      type: object
      required:
        [
          userId,
          email,
          joinedAt,
          periodStart,
          periodEnd,
          promptTokens,
          completionTokens,
          totalTokens,
          requestCount,
        ]
      properties:
        userId:
          type: integer
        email:
          type: string
          format: email
        name:
          type: ["string", "null"]
          description: Display name of the super admin, if set.
        joinedAt:
          type: string
          format: date-time
          description: The super admin's join date — anchors the billing period day-of-month.
        periodStart:
          type: string
          format: date-time
          description: Inclusive start of the current billing period.
        periodEnd:
          type: string
          format: date-time
          description: Exclusive end of the current billing period (start of the next).
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        totalTokens:
          type: integer
        requestCount:
          type: integer
          description: Number of AI completion calls in the period.
    PricingConfig:
      type: object
      required:
        [
          dbPricePer500Mb,
          userPricePerUser,
          channelPricePer2,
          aiPricePer100Tokens,
        ]
      properties:
        dbPricePer500Mb:
          type: integer
          description: Rupiah per 500 MB of chat storage per month.
        userPricePerUser:
          type: integer
          description: Rupiah per invited child user (supervisor/agent) per month.
        channelPricePer2:
          type: integer
          description: Rupiah per 2 channels per month.
        aiPricePer100Tokens:
          type: integer
          description: Rupiah per 100 AI tokens per month.
    UpdatePricingInput:
      type: object
      required:
        [
          dbPricePer500Mb,
          userPricePerUser,
          channelPricePer2,
          aiPricePer100Tokens,
        ]
      properties:
        dbPricePer500Mb:
          type: integer
          minimum: 0
        userPricePerUser:
          type: integer
          minimum: 0
        channelPricePer2:
          type: integer
          minimum: 0
        aiPricePer100Tokens:
          type: integer
          minimum: 0
...
    BillingUsage:
      type: object
      required:
        [storageBytes, mediaStorageBytes, childUserCount, channelCount, tokenUsage]
      properties:
        storageBytes:
          type: integer
          format: int64
          description: Chat-storage (DB) footprint in bytes — legacy metered metric.
        mediaStorageBytes:
          type: integer
          format: int64
          description: Object Storage footprint in bytes (SUM of media_objects) — measured against the storage quota.
        childUserCount:
          type: integer
          description: Number of invited child users (parent excluded).
        channelCount:
          type: integer
        tokenUsage:
          type: integer
          description: AI tokens consumed in the current billing period.
    BillBreakdown:
      type: object
      required: [dbCharge, userCharge, channelCharge, aiCharge, total]
      properties:
        dbCharge:
          type: integer
        userCharge:
          type: integer
        channelCharge:
          type: integer
        aiCharge:
          type: integer
        total:
          type: integer
    SubscriptionInfo:
      type: object
      required: [status, effectiveStatus, readOnly, currentPeriodEnd, createdAt]
      properties:
        status:
          type: string
          enum: [trial, active, expired, suspended]
          description: The stored subscription status.
        effectiveStatus:
          type: string
          enum: [trial, active, expired, suspended]
          description: >-
            Status computed live against the wall clock — an overdue trial/active
            collapses to "expired". This is what enforcement reacts to.
        readOnly:
          type: boolean
          description: >-
            True when the tenant is in read-only mode (effective status expired or
            suspended): viewing is allowed but every write is blocked.
        currentPeriodEnd:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    AdminTenantBilling:
      type: object
      required: [userId, email, status, storageLimit, usage, breakdown]
      properties:
        userId:
          type: integer
        email:
          type: string
          format: email
        name:
          type: ["string", "null"]
        status:
          type: string
          enum: [trial, active, expired, suspended]
        currentPeriodEnd:
          type: ["string", "null"]
          format: date-time
        storageLimit:
          type: integer
          format: int64
          description: Object Storage plafon in bytes (plan baseline + add-on top-ups).
        usage:
          $ref: "#/components/schemas/BillingUsage"
        breakdown:
          $ref: "#/components/schemas/BillBreakdown"
    RevenueTrendPoint:
      type: object
      required:
        [date, totalCharge, dbCharge, userCharge, channelCharge, aiCharge]
      properties:
        date:
          type: string
          description: Day in YYYY-MM-DD.
        totalCharge:
          type: integer
        dbCharge:
          type: integer
        userCharge:
          type: integer
        channelCharge:
          type: integer
        aiCharge:
          type: integer
...
    RevenueSummary:
      type: object
      required:
        - mrr
        - arr
        - arpu
        - totalTenants
        - activeTenants
        - trialTenants
        - expiredTenants
        - suspendedTenants
        - payingTenants
        - trend
      properties:
        mrr:
          type: integer
          description: Monthly recurring revenue — sum of latest snapshot total for effective-active owners.
        arr:
          type: integer
          description: Annual recurring revenue (mrr * 12).
        arpu:
          type: integer
          description: Average revenue per paying tenant (mrr / payingTenants).
        totalTenants:
          type: integer
        activeTenants:
          type: integer
        trialTenants:
          type: integer
        expiredTenants:
          type: integer
        suspendedTenants:
          type: integer
        payingTenants:
          type: integer
        trend:
          type: array
          description: Platform-wide daily total, oldest-first.
          items:
            $ref: "#/components/schemas/RevenueTrendPoint"
    RenewSubscriptionInput:
      type: object
      properties:
        status:
          type: string
          enum: [trial, active, expired, suspended]
          description: New stored status. "active" with extendMonths is the "mark paid" action.
        extendMonths:
          type: integer
          minimum: 1
          maximum: 24
          description: Extend the current period end by this many months.
        setUnlimited:
          type: boolean
          description: >-
            Grant infinite validity — forces status "active" and clears the
            period end so the tenant never expires. Takes precedence over
            status/extendMonths.
...