# lib/api-zod/src/generated/types/billingUsage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Maximizing Your Chat with AI API
 * OpenAPI spec version: 0.1.0
 */

export interface BillingUsage {
  /** Chat-storage (DB) footprint in bytes — legacy metered metric. */
  storageBytes: bigint;
  /** Object Storage footprint in bytes (SUM of media_objects) — measured against the storage quota. */
  mediaStorageBytes: bigint;
  /** Number of invited child users (parent excluded). */
  childUserCount: number;
  channelCount: number;
  /** AI tokens consumed in the current billing period. */
  tokenUsage: number;
}